[Feat] 일정 예약 보정 추가 구현

+++ caramel-domain/src/main/kotlin/com/whatever/caramel/domain/notification/service/ScheduledNotificationService.kt
@@ -3,7 +3,7 @@ package com.whatever.caramel.domain.notification.service
 import com.whatever.caramel.domain.notification.model.NotificationType
 import com.whatever.caramel.domain.notification.model.ScheduledNotification
 import com.whatever.caramel.domain.notification.repository.ScheduledNotificationRepository
-import com.whatever.caramel.domain.notification.vo.NotificationMessage
+import com.whatever.caramel.domain.notification.vo.NotificationMessageVo
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 import java.time.LocalDateTime
@@ -13,8 +13,7 @@ class ScheduledNotificationService(
     private val scheduledNotificationRepository: ScheduledNotificationRepository,
 ) {
     fun scheduleNotifications(
-        messagesByUserId: Map<Long, NotificationMessage>,
-        notificationType: NotificationType,
+        messagesByUserId: Map<Long, NotificationMessageVo>,
         notifyAt: LocalDateTime,
         image: String? = null,
     ) {
@@ -25,7 +24,7 @@ class ScheduledNotificationService(
         val notifications = messagesByUserId.map { (userId, message) ->
             ScheduledNotification(
                 targetUserId = userId,
-                notificationType = notificationType,
+                notificationType = message.type,
                 notifyAt = notifyAt,
                 title = message.title,
                 body = message.body,
@@ -44,6 +43,10 @@ class ScheduledNotificationService(
         targetUserIds: Set<Long>,
         notificationTypes: Set<NotificationType>,
     ): Int {
+        if (targetUserIds.isEmpty() || notificationTypes.isEmpty()) {
+            return 0
+        }
+
         return scheduledNotificationRepository.deleteAllByNotificationTypeInAndTargetUserIdIn(
             notificationTypes = notificationTypes,
             targetUserIds = targetUserIds
